이름및 구조 수정 끝
