+++ src/config/RouterConfig.js
@@ -2,9 +2,9 @@ import React from 'react';
 import HomeComponent from '../components/HomeComponent';
 import LoginPage from '../components/LoginPage/LoginPage';
 import ProjectPage from '../components/ProjectPage/ProjectPage';
-import TEST from '../components/search/main';
-import MATRIXS from '../components/search/Matrixs/MatrixsPage';
-import SYSTEM from '../components/search/system/SystemPage';
+import TEST from '../components/specificRunPage/main';
+import MATRIXS from '../components/specificRunPage/Matrixs/MatrixsPage';
+import SYSTEM from '../components/specificRunPage/system/SystemPage';
 import WorkSpaceComponent from '../components/WorkSpace/WorkSpaceComponent';
 import HPOList from '../components/WorkSpace/HPOPage';
 import HPOSpace from '../components/HPOSpace/hpopage';
